+++ src/main/java/shop/heartmuscle/heartmuscle/controller/FeedController.java
@@ -3,29 +3,24 @@ package shop.heartmuscle.heartmuscle.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.modelmapper.TypeToken;
 import org.springframework.data.domain.Page;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import shop.heartmuscle.heartmuscle.entity.Feed;
-import shop.heartmuscle.heartmuscle.entity.User;
-import shop.heartmuscle.heartmuscle.dto.*;
 import shop.heartmuscle.heartmuscle.dto.request.CommentRequestDto;
 import shop.heartmuscle.heartmuscle.dto.request.FeedRequestDto;
-import shop.heartmuscle.heartmuscle.dto.response.UserResponseDto;
 import shop.heartmuscle.heartmuscle.repository.UserRepository;
 import shop.heartmuscle.heartmuscle.security.UserDetailsImpl;
-import shop.heartmuscle.heartmuscle.service.FeedServiceImpl;
+import shop.heartmuscle.heartmuscle.service.FeedService;
 import shop.heartmuscle.heartmuscle.service.UserService;
 
 import java.io.IOException;
-import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
 public class FeedController {
 
-    private final FeedServiceImpl feedService;
+    private final FeedService feedService;
     private final UserRepository userRepository;
     private final UserService userService;
     private final ModelMapper modelMapper;
@@ -35,8 +30,6 @@ public class FeedController {
     @Operation(description = "피드 저장하기", method = "POST")
     @PostMapping("/feed")
     public void ImageFeed(FeedRequestDto feedRequestDto, @AuthenticationPrincipal UserDetailsImpl nowUser) throws IOException {
-        System.out.println("유저네임여기같이들어오는지보자" + feedRequestDto.getUsername());
-        System.out.println("피드이미지여기같이들어오는지보자" + feedRequestDto.getImage());
         feedService.saveFeed(feedRequestDto, nowUser);
     }
